# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "This is the OpenAPI specification for your project."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateProfessionalProfileRequestSchema": {
        "type": "object",
        "properties": {
          "bio": {
            "type": "string",
            "maxLength": 1000,
            "nullable": true
          },
          "yearsOfExperience": {
            "type": "number",
            "minimum": 0,
            "maximum": 70,
            "nullable": true
          },
          "certifications": {
            "type": "string",
            "nullable": true
          },
          "specialties": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "hourlyRate": {
            "type": "number",
            "minimum": 0,
            "nullable": true
          },
          "bankAccount": {
            "type": "string",
            "nullable": true
          },
          "taxId": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UpdateProfessionalProfileRequestSchema": {
        "type": "object",
        "properties": {
          "bio": {
            "type": "string",
            "maxLength": 1000,
            "nullable": true
          },
          "yearsOfExperience": {
            "type": "number",
            "minimum": 0,
            "maximum": 70,
            "nullable": true
          },
          "certifications": {
            "type": "string",
            "nullable": true
          },
          "specialties": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "hourlyRate": {
            "type": "number",
            "minimum": 0,
            "nullable": true
          },
          "bankAccount": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Authentication required"
                },
                "code": {
                  "type": "string",
                  "example": "401"
                }
              }
            }
          }
        }
      },
      "403": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Access denied"
                },
                "code": {
                  "type": "string",
                  "example": "403"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "code": {
                  "type": "string",
                  "example": "500"
                }
              }
            }
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/professionals/list": {
      "get": {
        "operationId": "get-professionals-list",
        "summary": "Get professionals list",
        "description": "Retrieve paginated list of professionals with optional filters",
        "tags": [
          "Professionals"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessionalsListResponseSchema:Professionals retrieved successfully"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/professionals/profile": {
      "get": {
        "operationId": "get-professionals-profile",
        "summary": "Get current professional profile",
        "description": "Retrieve the authenticated professional's profile information",
        "tags": [
          "Professionals"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessionalProfileResponseSchema:Professional profile retrieved successfully"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Professional profile not found"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/professionals/{id}/stats": {
      "get": {
        "operationId": "get-professionals-{id}-stats",
        "summary": "Get professional statistics",
        "description": "Retrieve statistics for a specific professional",
        "tags": [
          "Professionals"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessionalStatsResponseSchema:Professional statistics retrieved successfully"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invalid professional ID"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/professionals/profile/create": {
      "post": {
        "operationId": "post-professionals-profile-create",
        "summary": "Create professional profile",
        "description": "Create a professional profile for the authenticated user",
        "tags": [
          "Professionals"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfessionalProfileRequestSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessionalProfileResponseSchema:Professional profile created successfully"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Only clients and professionals can create a professional profile"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/professionals/profile/update": {
      "put": {
        "operationId": "put-professionals-profile-update",
        "summary": "Update professional profile",
        "description": "Update the authenticated professional's profile",
        "tags": [
          "Professionals"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfessionalProfileRequestSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessionalProfileResponseSchema:Professional profile updated successfully"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Professional profile not found"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  }
}